+++ packages/web/src/components/usdc-purchase-details-modal/components/DetailSection.tsx
@@ -1,0 +1,18 @@
+import { Text } from 'components/typography'
+
+import styles from './styles.module.css'
+
+export const DetailSection = ({
+  children,
+  label
+}: {
+  children: React.ReactNode
+  label: string
+}) => (
+  <div className={styles.detailSection}>
+    <Text variant='label' size='large' color='neutralLight4'>
+      {label}
+    </Text>
+    {children}
+  </div>
+)
