+++ packages/web/src/components/usdc-purchase-details-modal/components/types.ts
@@ -1,0 +1,6 @@
+import { USDCPurchaseDetails } from '@audius/common'
+
+export type ContentProps = {
+  purchaseDetails: USDCPurchaseDetails
+  onClose: () => void
+}
